+++ CHANGELOG.md
@@ -35,6 +35,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to have useless nodes
 - Forbids to have useless `raise` statements
 - Adds `params` and `parameters` to black-listed names
+- Forbids to use redundant `+` operator before numbers
 
 
 ### Bugfixes